requirements.txt
python-dotenv
chromadb
tiktoken
pypdf
langchain
streamlit
streamlit-chat
pinecone-client